generators.scala eliminated
